フォロー機能

+++ app/User.php
@@ -49,9 +49,19 @@ class User extends Authenticatable
         return $this->belongsToMany(User::class, 'follows', 'user_id', 'follow_user_id')->withTimestamps();
     }
 
+    public function follower()
+    {
+        return $this->belongsToMany(User::class, 'follows', 'follow_user_id', 'user_id')->withTimestamps();
+    }
+
     public function follow($userId)
     {
         $exist = $this->isFollow($userId);
+        //自分自身をフォローしない
+        if ($this->id === $userId) {
+            return false;
+        }
+
         if ($exist) {
             return false;
         } else {
@@ -63,6 +73,11 @@ class User extends Authenticatable
     public function unfollow($userId)
     {
         $exist = $this->isFollow($userId);
+        //自分自身をフォローしない
+        if ($this->id === $userId) {
+            return false;
+        }
+
         if ($exist) {
             $this->following()->detach($userId);
             return true;
@@ -74,11 +89,6 @@ class User extends Authenticatable
     public function isFollow($userId)
     {
         return $this->following()->where('follow_user_id', $userId)->exists();
-    }
-
-    public function follower()
-    {
-        return $this->belongsToMany(User::class, 'follows', 'follow_user_id', 'user_id')->withTimestamps();
     }
 
     public static function boot()
